+++ diesel/src/types/impls/option.rs
@@ -56,6 +56,26 @@ impl<T, ST, DB> Queryable<Nullable<ST>, DB> for Option<T> where
     }
 }
 
+impl<T, ST, DB> FromSqlRow<Nullable<ST>, DB> for Option<T> where
+    T: FromSqlRow<ST, DB>,
+    DB: Backend + HasSqlType<ST>,
+    ST: NotNull,
+{
+    fn build_from_row<R: ::row::Row<DB>>(row: &mut R) -> Result<Self, Box<Error+Send+Sync>> {
+        let fields_needed = Self::fields_needed();
+        if row.next_is_null(fields_needed) {
+            row.advance(fields_needed);
+            Ok(None)
+        } else {
+            T::build_from_row(row).map(Some)
+        }
+    }
+
+    fn fields_needed() -> usize {
+        T::fields_needed()
+    }
+}
+
 impl<T, ST, DB> ToSql<Nullable<ST>, DB> for Option<T> where
     T: ToSql<ST, DB>,
     DB: Backend + HasSqlType<ST>,
